Update types, remove blog module from page
